chore: Add wifi command to shellAliases in default.nix
This commit adds the "wifi" command to the shellAliases in the default.nix file. The "wifi" command is used to launch the "nmtui" tool, which provides a text-based user interface for managing network connections. This change enhances the functionality of the shell by providing a convenient way to access the network management tool.