refactor: Update zathura.nix to include font configuration
This commit updates the zathura.nix file to include font configuration using the `config.stylix.fonts` property. The `font` option is set to `${monospace.name}` to use the monospace font defined in the configuration. Additionally, the `recolor` option is set to `true` to enable book content color recoloring. These changes enhance the visual experience and customization options of the Zathura application.