feat: Add custom/applauncher module to waybar configuration
This commit adds the "custom/applauncher" module to the waybar configuration. The module includes a custom application launcher with the format "󱗼" and a tooltip format of "Application Launcher". Clicking on the launcher opens the "applauncher-fullscreen.sh" script using the xdg config home path. This enhancement improves the user experience by providing a convenient and customizable application launcher option in the waybar menu.