feat: Update global locale configuration
This commit updates the global locale configuration in the `locale.nix` file. It modifies the `services.geoclue2` and `services.timesyncd` sections, enabling the `geoclue2` service and disabling data submission, and enabling the `timesyncd` service. Additionally, it updates the `location.provider` to use `geoclue2` and sets the `time.hardwareClockInLocalTime` to `false` and `time.timeZone` to `America/Chicago`. These changes enhance the functionality and user experience of the application by improving location and time synchronization.