chore: Add dunst to exec-once in default.nix
This commit adds the `dunst` executable to the `exec-once` section in the `default.nix` file. The `dunst` package is obtained using the `lib.getExe` function from the `config.services.dunst.package` configuration. This change enhances the functionality of the hyprland desktop feature by including the `dunst` notification daemon in the startup processes.