chore: Update waybar colors and opacity
This commit updates the colors and opacity of the waybar in the `default.nix` file. The `textcolor1` is changed from `@base0E` to `@base07`, and the `background-color` is changed from `@backgrounddark2` to `@backgrounddark1`. Additionally, the opacity is increased from 0.7 to 0.8. These changes improve the visual appearance of the waybar in the hyprland desktop feature.