Add nautilus configuration to common optional hosts
This commit adds a new configuration file, nautilus.nix, to the common optional hosts directory. The configuration enables the nautilus-open-any-terminal program and sets the terminal to "kitty". It also adds system packages gnome.nautilus and gnome.nautilus-python to the environment. Additionally, it sets the session variable NAUTILUS_4_EXTENSION_DIR to the appropriate path and adds a path to link for nautilus-python extensions.

Modified files:
- hosts/common/optional/nautilus.nix
- hosts/frametop/default.nix
- hosts/woody/default.nix